# backend/swagger.yaml
openapi: 3.0.0
info:
  title: MSE - Main Solution Estates
  description: MongoDB ExpressJS NodeJS REST API for a Real Estate App.<br />Github repository available [here](https://github.com/BenjaminKing1337/fullstack_project-MSE).<br />Some endpoints require `auth-token` in order to be executed.
  version: 1.0.0
servers:
  - url: https://mse-a9ir.onrender.com/api/
    description: Remote server
  - url: http://localhost:3000/api/
    description: Localhost development server
components:
  securitySchemes:
    ApiKeyAuth: 
      type: apiKey
      in: header 
      name: auth-token
  schemas:
...
    Customer:
      type: object
      properties:
        forename:
          type: string
        surname:
          type: string
        email:
          type: string
        company_name:
          type: string
        contact_number:
          type: string
        user_id:
          type: string
        created_by:
          type: string
          description: "Id of the logged in user"
        created_at:
          type: string
          description: "It takes the current date as default"
          example: "2014-01-01T23:28:56.782Z"
...
    Document:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              is_active:
                type: boolean
        property_id:
          type: string
        building_id:
          type: string
        created_by:
          type: string
          description: "Id of the logged in user"
        created_at:
          type: string
          description: "It takes the current date as default"
          example: "2014-01-01T23:28:56.782Z"
paths:
...
  /customers:
    get:
      summary: Get all customers
      tags:
        - Customer
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Customers successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /customers/get/byUser/{id}:
    get:
      summary: Get customers by logged in user ID
      tags:
        - Customer
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Customers successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /customers/get/{id}:
    get:
      summary: Get customer by ID
      tags:
        - Customer
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Customer successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /customers/new:
    post:
      summary: New customer
      tags:
        - Customer
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Customer"
      responses:
        200:
          description: Customer successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /customers/update/{id}:
    put:
      summary: Update customer by ID
      tags:
        - Customer
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Customer"
      responses:
        200:
          description: Customer successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /customers/delete/{id}:
    delete:
      summary: Delete customer by ID
      tags:
        - Customer
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Customer successfully deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /documents:
    get:
      summary: Get all documents
      tags:
        - Document
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Documents successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /documents/get/byUser/{id}:
    get:
      summary: Get documents by logged in user ID
      tags:
        - Document
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Documents successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /documents/get/{id}:
    get:
      summary: Get document by ID
      tags:
        - Document
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Document successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /documents/new:
    post:
      summary: New document
      tags:
        - Document
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Document"
      responses:
        200:
          description: Document successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /documents/update/{id}:
    put:
      summary: Update document by ID
      tags:
        - Document
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Document"
      responses:
        200:
          description: Document successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /documents/delete/{id}:
    delete:
      summary: Delete document by ID
      tags:
        - Document
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Document successfully deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string